requirements.txt
anyio==3.6.1
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
firebase-admin==5.2.0
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-appengine-logging==1.1.2
google-cloud-audit-log==0.2.2
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-logging==3.2.1
google-cloud-pubsub==2.13.4
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.13.0
httplib2==0.20.4
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
orjson==3.7.7
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
ujson==5.4.0
uritemplate==4.1.1
urllib3==1.26.10
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
Werkzeug==2.1.2
zipp==3.8.1
